RecoverEEE can now have multiple recovery targets

+++ src/states/Recovery.zig
@@ -30,15 +30,13 @@ const files = system.files;
 
 const strings = sandeee_data.strings;
 
-const VERSION = "seed#1_0";
+const VERSION = "seed#2_0";
 
 const GSRecovery = @This();
 
-const RecoveryMenuEntry = enum {
-    Reinstall,
-    ReinstallReset,
-    Delete,
-    Back,
+const RecoveryChoice = struct {
+    name: []const u8,
+    disk: []const u8 = "",
 };
 
 shader: *Shader,
@@ -50,8 +48,9 @@ select_sound: *audio.Sound,
 sel: usize = 0,
 disks: std.array_list.Managed([]const u8) = .init(allocator),
 status: []const u8 = "",
-sub_sel: ?RecoveryMenuEntry = null,
+sub_sel: ?usize = null,
 confirm_sel: ?bool = null,
+recovery_choices: std.array_list.Managed(RecoveryChoice) = .init(allocator),
 
 const DISK_LIST = "0123456789ABCDEF";
 const TEXT_COLOR = Color{ .r = 1, .g = 1, .b = 1 };
@@ -65,6 +64,10 @@ pub fn getDate(name: []const u8) i128 {
     return (file.stat() catch return 0).mtime;
 }
 
+pub fn sortRecoveryLt(_: void, a: RecoveryChoice, b: RecoveryChoice) bool {
+    return std.mem.lessThan(u8, a.name, b.name);
+}
+
 pub fn sortDisksLt(_: void, a: []const u8, b: []const u8) bool {
     return getDate(a) < getDate(b);
 }
@@ -76,49 +79,108 @@ pub fn setup(self: *GSRecovery) !void {
     self.sub_sel = null;
     self.status = "";
 
-    var dir = try std.fs.cwd().openDir("disks", .{
-        .iterate = true,
-    });
-    defer dir.close();
+    // setup disk list
+    {
+        var dir = try std.fs.cwd().openDir("disks", .{
+            .iterate = true,
+        });
+        defer dir.close();
 
-    var iter = dir.iterate();
+        var iter = dir.iterate();
 
-    while (try iter.next()) |item| {
-        if (!std.mem.endsWith(u8, item.name, ".eee")) continue;
+        while (try iter.next()) |item| {
+            if (!std.mem.endsWith(u8, item.name, ".eee")) continue;
 
-        const entry = try allocator.dupe(u8, item.name);
+            const entry = try allocator.dupe(u8, item.name);
 
-        try self.disks.append(entry);
+            try self.disks.append(entry);
+        }
+
+        std.sort.insertion([]const u8, self.disks.items, {}, sortDisksLt);
+
+        for (self.disks.items, 0..) |_, idx| {
+            const copy = self.disks.items[idx];
+            defer allocator.free(copy);
+
+            self.disks.items[idx] = try std.fmt.allocPrint(allocator, "{c} {s}", .{ DISK_LIST[idx], copy });
+        }
+
+        self.disks.items.len = @min(self.disks.items.len, DISK_LIST.len);
+
+        try self.disks.append("X Back");
     }
 
-    std.sort.insertion([]const u8, self.disks.items, {}, sortDisksLt);
+    // setup recovery option list
+    {
+        var dir = try std.fs.cwd().openDir("content", .{
+            .iterate = true,
+        });
+        defer dir.close();
 
-    for (self.disks.items, 0..) |_, idx| {
-        const copy = self.disks.items[idx];
-        defer allocator.free(copy);
+        var iter = dir.iterate();
 
-        self.disks.items[idx] = try std.fmt.allocPrint(allocator, "{c} {s}", .{ DISK_LIST[idx], copy });
+        while (try iter.next()) |item| {
+            if (!std.mem.endsWith(u8, item.name, ".eee")) continue;
+
+            var file = try dir.openFile(item.name, .{ .mode = .read_only });
+            defer file.close();
+
+            const disk = try allocator.dupe(u8, item.name);
+            var name = try allocator.dupe(u8, item.name);
+
+            const conts = try files.Folder.diskRecoveryInfo(file);
+            if (conts) |c| {
+                var split = std.mem.splitScalar(u8, c, '\n');
+
+                while (split.next()) |entry| {
+                    const line = std.mem.trim(u8, entry, &std.ascii.whitespace);
+                    if (std.mem.containsAtLeast(u8, line, 2, "=")) continue;
+
+                    const equals_idx = std.mem.indexOf(u8, line, "=") orelse continue;
+
+                    const menu_entry = std.mem.trim(u8, line[0..equals_idx], &std.ascii.whitespace);
+                    if (std.mem.eql(u8, menu_entry, "menu_option")) {
+                        const assigns = std.mem.trim(u8, line[equals_idx + 1 ..], &std.ascii.whitespace);
+
+                        name = try allocator.realloc(name, assigns.len);
+                        @memcpy(name, assigns);
+                    }
+                }
+            }
+            defer if (conts) |c| allocator.free(c);
+
+            try self.recovery_choices.append(.{
+                .name = name,
+                .disk = disk,
+            });
+        }
+
+        std.sort.insertion(RecoveryChoice, self.recovery_choices.items, {}, sortRecoveryLt);
+
+        for (self.recovery_choices.items, 0..) |_, idx| {
+            const copy = self.recovery_choices.items[idx].name;
+            defer allocator.free(copy);
+
+            self.recovery_choices.items[idx].name = try std.fmt.allocPrint(allocator, "{c} {s}", .{ DISK_LIST[idx], copy });
+        }
+
+        try self.recovery_choices.append(.{ .name = "D Delete Disk" });
+        try self.recovery_choices.append(.{ .name = "X Back" });
     }
-
-    self.disks.items.len = @min(self.disks.items.len, DISK_LIST.len);
-
-    try self.disks.append("X Back");
 }
 
 pub fn deinit(self: *GSRecovery) void {
     for (self.disks.items[0 .. self.disks.items.len - 1]) |item| {
         allocator.free(item);
     }
-
     self.disks.clearAndFree();
-}
 
-const UPDATE_MODES = [_][*:0]const u8{
-    "R Reinstall System Files",
-    "S Reinstall System Files and Default Settings",
-    "D Delete disk",
-    "X Back",
-};
+    for (self.recovery_choices.items[0 .. self.recovery_choices.items.len - 2]) |item| {
+        allocator.free(item.name);
+        allocator.free(item.disk);
+    }
+    self.recovery_choices.clearAndFree();
+}
 
 const CONFIRM = [_][*:0]const u8{
     "Y Yes",
@@ -147,12 +209,13 @@ pub fn draw(self: *GSRecovery, _: Vec2) !void {
     y += self.face.size * 2;
 
     if (self.confirm_sel) |confirm_sel| {
-        const prompt = switch (self.sub_sel orelse unreachable) {
-            .Reinstall => "Reinstall All System Files?",
-            .ReinstallReset => "Reinstall All System Files And Settings?",
-            .Delete => "Delete this disk?",
-            else => "",
-        };
+        const prompt = if (self.sub_sel) |sub_sel|
+            if (sub_sel == self.recovery_choices.items.len - 2)
+                "Delete this disk?"
+            else
+                "Reinstall all system files?"
+        else
+            unreachable;
 
         try self.face.draw(.{
             .shader = self.font_shader,
@@ -183,11 +246,11 @@ pub fn draw(self: *GSRecovery, _: Vec2) !void {
             y += self.face.size * 1;
         }
     } else if (self.sub_sel) |sub_sel| {
-        for (UPDATE_MODES, 0..) |mode, idx| {
-            const line = try std.fmt.allocPrint(allocator, "  {s}", .{mode});
+        for (self.recovery_choices.items, 0..) |choice, idx| {
+            const line = try std.fmt.allocPrint(allocator, "  {s}", .{choice.name});
             defer allocator.free(line);
 
-            if (idx == @intFromEnum(sub_sel)) {
+            if (idx == sub_sel) {
                 line[0] = '>';
             }
 
@@ -237,42 +300,33 @@ pub fn keypress(self: *GSRecovery, key: c_int, _: c_int, down: bool) !void {
             if (self.sub_sel) |sub_sel| {
                 if (self.confirm_sel) |confirm| {
                     if (confirm) {
-                        switch (sub_sel) {
-                            .Reinstall => {
-                                try files.Folder.recoverDisk(self.disks.items[self.sel][2..], false);
-                                self.status = "Reinstalled";
+                        if (sub_sel == self.recovery_choices.items.len - 2) {
+                            const path = try std.fmt.allocPrint(allocator, "disks/{s}", .{self.disks.items[self.sel][2..]});
+                            defer allocator.free(path);
+
+                            self.status = "Deleted";
+
+                            try std.fs.cwd().deleteFile(path);
+
+                            _ = self.disks.orderedRemove(self.sel);
+
+                            self.confirm_sel = null;
+                            self.sub_sel = null;
+                            self.sel = 0;
+
+                            if (self.disks.items.len == 1) {
+                                try events.EventManager.instance.sendEvent(system_events.EventStateChange{
+                                    .target_state = .Disks,
+                                });
+
                                 try audio.instance.playSound(self.select_sound.*);
-                            },
-                            .ReinstallReset => {
-                                try files.Folder.recoverDisk(self.disks.items[self.sel][2..], true);
-                                self.status = "Reinstalled & Reset";
-                                try audio.instance.playSound(self.select_sound.*);
-                            },
-                            .Delete => {
-                                const path = try std.fmt.allocPrint(allocator, "disks/{s}", .{self.disks.items[self.sel][2..]});
-                                defer allocator.free(path);
+                            }
 
-                                self.status = "Deleted";
-
-                                try std.fs.cwd().deleteFile(path);
-
-                                _ = self.disks.orderedRemove(self.sel);
-
-                                self.confirm_sel = null;
-                                self.sub_sel = null;
-                                self.sel = 0;
-
-                                if (self.disks.items.len == 1) {
-                                    try events.EventManager.instance.sendEvent(system_events.EventStateChange{
-                                        .target_state = .Disks,
-                                    });
-
-                                    try audio.instance.playSound(self.select_sound.*);
-                                }
-
-                                return;
-                            },
-                            else => {},
+                            return;
+                        } else if (sub_sel < self.recovery_choices.items.len - 2) {
+                            try files.Folder.recoverDisk(self.disks.items[self.sel][2..], self.recovery_choices.items[sub_sel].disk, false);
+                            self.status = "Reinstalled";
+                            try audio.instance.playSound(self.select_sound.*);
                         }
                     }
 
@@ -280,11 +334,11 @@ pub fn keypress(self: *GSRecovery, key: c_int, _: c_int, down: bool) !void {
 
                     try audio.instance.playSound(self.select_sound.*);
 
-                    self.sub_sel = .Reinstall;
+                    self.sub_sel = 0;
                     self.sel = 0;
 
                     return;
-                } else if (sub_sel == .Back) {
+                } else if (sub_sel == self.recovery_choices.items.len - 1) {
                     self.sub_sel = null;
                     try audio.instance.playSound(self.select_sound.*);
                 } else {
@@ -298,7 +352,7 @@ pub fn keypress(self: *GSRecovery, key: c_int, _: c_int, down: bool) !void {
 
                 try audio.instance.playSound(self.select_sound.*);
             } else {
-                self.sub_sel = .Reinstall;
+                self.sub_sel = 0;
 
                 try audio.instance.playSound(self.select_sound.*);
             }
@@ -307,9 +361,8 @@ pub fn keypress(self: *GSRecovery, key: c_int, _: c_int, down: bool) !void {
             if (self.confirm_sel != null) {
                 self.confirm_sel = false;
             } else if (self.sub_sel) |*sub_sel| {
-                if (@intFromEnum(sub_sel.*) < @intFromEnum(RecoveryMenuEntry.Back)) {
-                    sub_sel.* = @enumFromInt(@intFromEnum(sub_sel.*) + 1);
-
+                if (sub_sel.* < self.recovery_choices.items.len - 1) {
+                    sub_sel.* += 1;
                     try audio.instance.playSound(self.blip_sound.*);
                 }
             } else if (self.sel < self.disks.items.len - 1) {
@@ -321,8 +374,8 @@ pub fn keypress(self: *GSRecovery, key: c_int, _: c_int, down: bool) !void {
             if (self.confirm_sel != null) {
                 self.confirm_sel = true;
             } else if (self.sub_sel) |*sub_sel| {
-                if (@intFromEnum(sub_sel.*) > @intFromEnum(RecoveryMenuEntry.Reinstall)) {
-                    sub_sel.* = @enumFromInt(@intFromEnum(sub_sel.*) - 1);
+                if (sub_sel.* != 0) {
+                    sub_sel.* -= 1;
                     try audio.instance.playSound(self.blip_sound.*);
                 }
             } else if (self.sel != 0) {
@@ -347,16 +400,6 @@ pub fn keypress(self: *GSRecovery, key: c_int, _: c_int, down: bool) !void {
             } else if (self.sub_sel) |sub_sel| {
                 _ = sub_sel;
                 switch (std.ascii.toUpper(name[0])) {
-                    'R' => {
-                        try files.Folder.recoverDisk(self.disks.items[self.sel][2..], false);
-                        self.status = "Reinstalled";
-                        try audio.instance.playSound(self.select_sound.*);
-                    },
-                    'S' => {
-                        try files.Folder.recoverDisk(self.disks.items[self.sel][2..], true);
-                        self.status = "Reinstalled & Reset";
-                        try audio.instance.playSound(self.select_sound.*);
-                    },
                     'D' => {
                         const path = try std.fmt.allocPrint(allocator, "disks/{s}", .{self.disks.items[self.sel][2..]});
                         defer allocator.free(path);
@@ -388,7 +431,7 @@ pub fn keypress(self: *GSRecovery, key: c_int, _: c_int, down: bool) !void {
                 for (self.disks.items, 0..) |disk, idx| {
                     if (std.ascii.toUpper(name[0]) == disk[0]) {
                         self.sel = idx;
-                        self.sub_sel = .Reinstall;
+                        self.sub_sel = 0;
 
                         try audio.instance.playSound(self.select_sound.*);
                     }
